fatfs: ccc

+++ src/fs/fat/ff.c
@@ -2120,8 +2120,8 @@ static WORD xname_sum (	/* Get check sum (to be used as hash) of the file name *
 
 	while ((chr = *name++) != 0) {
 		chr = (WCHAR)ff_wtoupper(chr);		/* File name needs to be up-case converted */
-		sum = ((sum & 1u) ? 0x8000u : 0) + (sum >> 1) + (chr & 0xFFu);
-		sum = ((sum & 1u) ? 0x8000u : 0) + (sum >> 1) + (chr >> 8);
+		sum = (WORD)(((sum & 1u) ? 0x8000u : 0) + (sum >> 1) + (chr & 0xFFu));
+		sum = (WORD)(((sum & 1u) ? 0x8000u : 0) + (sum >> 1) + (chr >> 8));
 	}
 	return sum;
 }
@@ -2513,7 +2513,7 @@ static FRESULT dir_register (	/* FR_OK:succeeded, FR_DENIED:no free entry or too
 		dp->blk_ofs = dp->dptr - SZDIRE * (n_ent - 1);	/* Set the allocated entry block offset */
 
 		if (dp->obj.stat & 4) {			/* Has the directory been stretched by new allocation? */
-			dp->obj.stat &= ~4;
+			dp->obj.stat &= ~4u;
 			res = fill_first_frag(&dp->obj);	/* Fill the first fragment on the FAT if needed */
 			if (res != FR_OK) return res;
 			res = fill_last_frag(&dp->obj, dp->clust, 0xFFFFFFFF);	/* Fill the last fragment on the FAT if needed */
